Allow SYNC commands to be triggered from button presses

+++ src/cli/pytrain.py
@@ -93,6 +93,7 @@ class PyTrain:
         self._force_update = False
         self._force_upgrade = False
         self._force_shutdown = False
+        self._xxx = set()
         self._script_loader: StartupScriptLoader | None = None
 
         if args.base is not None:
@@ -221,30 +222,35 @@ class PyTrain:
         """
         if self._echo:
             log.info(f"{datetime.now().strftime('%H:%M:%S.%f')[:-3]} {cmd}")
-        if self.is_client and cmd.command == TMCC1SyncCommandDef.QUIT:
-            log.info("Server exiting...")
-            # send keyboard interrupt to main process to shut ii down
-            os.kill(os.getpid(), signal.SIGINT)
-        elif cmd.command == TMCC1SyncCommandDef.REBOOT:
-            log.info(f"{'Server' if self.is_server else 'Client'} rebooting...")
-            self._force_reboot = True
-            os.kill(os.getpid(), signal.SIGINT)
-        elif cmd.command == TMCC1SyncCommandDef.RESTART:
-            log.info(f"{'Server' if self.is_server else 'Client'} restarting...")
-            self._force_restart = True
-            os.kill(os.getpid(), signal.SIGINT)
-        elif cmd.command == TMCC1SyncCommandDef.SHUTDOWN:
-            log.info(f"{'Server' if self.is_server else 'Client'} shutting down...")
-            self._force_shutdown = True
-            os.kill(os.getpid(), signal.SIGINT)
-        elif cmd.command == TMCC1SyncCommandDef.UPDATE:
-            log.info(f"{'Server' if self.is_server else 'Client'} updating...")
-            self._force_update = True
-            os.kill(os.getpid(), signal.SIGINT)
-        elif cmd.command == TMCC1SyncCommandDef.UPGRADE:
-            log.info(f"{'Server' if self.is_server else 'Client'} upgrading...")
-            self._force_upgrade = True
-            os.kill(os.getpid(), signal.SIGINT)
+
+        if cmd not in self._xxx:
+            self._xxx.add(cmd)
+            if self.is_client and cmd.command == TMCC1SyncCommandDef.QUIT:
+                log.info("Server exiting...")
+                # send keyboard interrupt to main process to shut ii down
+                os.kill(os.getpid(), signal.SIGINT)
+            elif cmd.command == TMCC1SyncCommandDef.REBOOT:
+                log.info(f"{'Server' if self.is_server else 'Client'} rebooting...")
+                self._force_reboot = True
+                os.kill(os.getpid(), signal.SIGINT)
+            elif cmd.command == TMCC1SyncCommandDef.RESTART:
+                log.info(f"{'Server' if self.is_server else 'Client'} restarting...")
+                self._force_restart = True
+                os.kill(os.getpid(), signal.SIGINT)
+            elif cmd.command == TMCC1SyncCommandDef.SHUTDOWN:
+                log.info(f"{'Server' if self.is_server else 'Client'} shutting down...")
+                self._force_shutdown = True
+                os.kill(os.getpid(), signal.SIGINT)
+            elif cmd.command == TMCC1SyncCommandDef.UPDATE:
+                log.info(f"{'Server' if self.is_server else 'Client'} updating...")
+                self._force_update = True
+                os.kill(os.getpid(), signal.SIGINT)
+            elif cmd.command == TMCC1SyncCommandDef.UPGRADE:
+                log.info(f"{'Server' if self.is_server else 'Client'} upgrading...")
+                self._force_upgrade = True
+                os.kill(os.getpid(), signal.SIGINT)
+        else:
+            print(f"Command '{cmd}' duplicated.")
 
     def __repr__(self) -> str:
         sc = "Server" if self.is_server else "Client"
